Refactor ServiceCollectionExtensions in HomeInventory.Web
