+++ src/HomeInventory/HomeInventory.Web/ServiceCollectionExtensions.cs
@@ -24,6 +24,7 @@ using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Swashbuckle.AspNetCore.SwaggerUI;
 
 namespace HomeInventory.Web;
 
@@ -43,52 +44,18 @@ public static class ServiceCollectionExtensions
         services.AddScoped<ICorrelationIdContainer, CorrelationIdContainer>();
         services.AddScoped<CorrelationIdMiddleware>();
 
+        services.AddMappingAssemblySource(AssemblyReference.Assembly);
         services.AddAutoMapper((sp, configExpression) =>
         {
             configExpression.AddMaps(sp.GetServices<IMappingAssemblySource>().Select(s => s.GetAssembly()));
             configExpression.ConstructServicesUsing(sp.GetService);
         }, Type.EmptyTypes);
 
-        services.AddMappingAssemblySource(AssemblyReference.Assembly);
+        services.AddAuthorization();
 
-        services.AddDynamicAuthorization();
+        services.AddOpenApiDocs();
 
-        services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-            .AddJwtBearer();
-
-        // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-        services.AddEndpointsApiExplorer();
-        services.ConfigureOptions<ConfigureSwaggerOptions>();
-        services.AddSingleton<IOpenApiValueConverter, JsonOpenApiValueConverter>();
-        services.AddSingleton<ISwaggerOperationFilter, DeprecatedSwaggerOperationFilter>();
-        services.AddSingleton<ISwaggerOperationFilter, ResponsesSwaggerOperationFilter>();
-        services.AddSingleton<ISwaggerOperationFilter, ParametersSwaggerOperationFilter>();
-        services.AddSwaggerGen(options =>
-            options.OperationFilter<SwaggerDefaultValues>());
-        services.AddApiVersioning(options =>
-        {
-            options.DefaultApiVersion = new ApiVersion(1);
-            options.AssumeDefaultVersionWhenUnspecified = true;
-            options.ApiVersionReader = new QueryStringApiVersionReader();
-        }).AddApiExplorer(options =>
-        {
-            options.GroupNameFormat = "'v'VVV";
-        });
-
-        services.AddSingleton<IJwtIdentityGenerator, GuidJwtIdentityGenerator>();
-        services.AddScoped<IAuthenticationTokenGenerator, JwtTokenGenerator>();
-
-        services.ConfigureOptions<JwtBearerOptionsSetup>();
-
-        services.AddFluentValidationAutoValidation(c =>
-        {
-            c.DisableDataAnnotationsValidation = true;
-        });
-        services.AddValidatorsFromAssembly(Contracts.Validations.AssemblyReference.Assembly, includeInternalTypes: true);
-        services.AddValidatorsFromAssembly(AssemblyReference.Assembly, ServiceLifetime.Singleton, filter: r => r.Is<IOptionsValidator>(), includeInternalTypes: true);
-        services.AddValidatorsFromAssembly(AssemblyReference.Assembly, ServiceLifetime.Scoped, filter: r => !r.Is<IOptionsValidator>(), includeInternalTypes: true);
-
-        services.AddOptionsWithValidator<JwtOptions>();
+        services.AddValidation();
 
         services.AddCarter(configurator: config => config
             .WithModule<UserManagementModule>()
@@ -98,35 +65,57 @@ public static class ServiceCollectionExtensions
         return services;
     }
 
+    private static void AddAuthorization(this IServiceCollection services)
+    {
+        services.AddDynamicAuthorization();
+
+        services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+            .AddJwtBearer();
+        services.ConfigureOptions<JwtBearerOptionsSetup>();
+
+        services.AddSingleton<IJwtIdentityGenerator, GuidJwtIdentityGenerator>();
+        services.AddScoped<IAuthenticationTokenGenerator, JwtTokenGenerator>();
+    }
+
+    private static void AddOpenApiDocs(this IServiceCollection services)
+    {
+        // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+        services.AddEndpointsApiExplorer();
+        services.ConfigureOptions<ConfigureSwaggerOptions>();
+        services.AddSingleton<IOpenApiValueConverter, JsonOpenApiValueConverter>();
+        services.AddSingleton<ISwaggerOperationFilter, DeprecatedSwaggerOperationFilter>();
+        services.AddSingleton<ISwaggerOperationFilter, ResponsesSwaggerOperationFilter>();
+        services.AddSingleton<ISwaggerOperationFilter, ParametersSwaggerOperationFilter>();
+        services.AddSwaggerGen(options => options.OperationFilter<SwaggerDefaultValues>());
+        services.AddApiVersioning(options =>
+        {
+            options.DefaultApiVersion = new ApiVersion(1);
+            options.AssumeDefaultVersionWhenUnspecified = true;
+            options.ApiVersionReader = new QueryStringApiVersionReader();
+        }).AddApiExplorer(options => options.GroupNameFormat = "'v'VVV");
+    }
+
+    private static void AddValidation(this IServiceCollection services)
+    {
+        services.AddFluentValidationAutoValidation(config => config.DisableDataAnnotationsValidation = true);
+        services.AddValidatorsFromAssembly(Contracts.Validations.AssemblyReference.Assembly, includeInternalTypes: true);
+        services.AddValidatorsFromAssembly(AssemblyReference.Assembly, ServiceLifetime.Singleton, filter: r => r.Is<IOptionsValidator>(), includeInternalTypes: true);
+        services.AddValidatorsFromAssembly(AssemblyReference.Assembly, ServiceLifetime.Scoped, filter: r => !r.Is<IOptionsValidator>(), includeInternalTypes: true);
+
+        services.AddOptionsWithValidator<JwtOptions>();
+    }
+
     public static TApp UseWeb<TApp>(this TApp app)
         where TApp : IApplicationBuilder, IEndpointRouteBuilder
     {
         app.UseSwagger();
-        app.UseSwaggerUI(options =>
-        {
-            var descriptions = app.DescribeApiVersions();
-            foreach (var description in descriptions)
-            {
-                AddSwaggerEndpoint(options, description);
-            }
-
-            static void AddSwaggerEndpoint(Swashbuckle.AspNetCore.SwaggerUI.SwaggerUIOptions options, Asp.Versioning.ApiExplorer.ApiVersionDescription description)
-            {
-                var url = $"/swagger/{description.GroupName}/swagger.json";
-                var name = description.GroupName.ToUpperInvariant();
-                options.SwaggerEndpoint(url, name);
-            }
-        });
+        app.UseSwaggerUI(app.ConfigureSwaggerUI);
 
         app.MapHealthChecks();
         app.UseHealthChecksUI();
 
         app.UseExceptionHandler(new ExceptionHandlerOptions { ExceptionHandlingPath = "/error", });
-        app.Map("/error", (HttpContext context) =>
-        {
-            var exception = context.GetFeature<IExceptionHandlerFeature>()?.Error;
-            return Results.Problem(detail: exception?.Message);
-        });
+        app.Map("/error", (HttpContext context) => Results.Problem(detail: context.GetFeature<IExceptionHandlerFeature>()?.Error?.Message));
 
         app.UseMiddleware<CorrelationIdMiddleware>();
 
@@ -138,6 +127,22 @@ public static class ServiceCollectionExtensions
         app.MapCarter();
 
         return app;
+    }
+
+    private static void ConfigureSwaggerUI(this IEndpointRouteBuilder builder, SwaggerUIOptions options)
+    {
+        var descriptions = builder.DescribeApiVersions();
+        foreach (var description in descriptions)
+        {
+            AddSwaggerEndpoint(options, description.GroupName);
+        }
+    }
+
+    private static void AddSwaggerEndpoint(SwaggerUIOptions options, string groupName)
+    {
+        var url = $"/swagger/{groupName}/swagger.json";
+        var name = groupName.ToUpperInvariant();
+        options.SwaggerEndpoint(url, name);
     }
 
     private static void MapHealthChecks(this IEndpointRouteBuilder app)
@@ -173,44 +178,5 @@ public static class ServiceCollectionExtensions
         var services = builder.Services;
         services.AddSingleton<IValidateOptions<TOptions>>(sp => new FluentOptionsValidator<TOptions>(builder.Name, sp.GetRequiredService<IValidator<TOptions>>(), validationOptions));
         return builder;
-    }
-
-    private sealed class FluentOptionsValidator<TOptions> : IValidateOptions<TOptions>
-        where TOptions : class
-    {
-        private readonly Action<ValidationStrategy<TOptions>>? _validationOptions;
-
-        public FluentOptionsValidator(string? name, IValidator<TOptions> validator, Action<ValidationStrategy<TOptions>>? validationOptions)
-        {
-            Name = name;
-            Validator = validator;
-            _validationOptions = validationOptions;
-        }
-
-        public string? Name { get; }
-
-        public IValidator<TOptions> Validator { get; }
-
-        public ValidateOptionsResult Validate(string? name, TOptions options)
-        {
-            // null name is used to configure all named options
-            if (Name != null && name != Name)
-            {
-                // ignored if not validating this instance
-                return ValidateOptionsResult.Skip;
-            }
-
-            var result = _validationOptions is null
-                ? Validator.Validate(options)
-                : Validator.Validate(options, _validationOptions);
-
-
-            if (result.IsValid)
-            {
-                return ValidateOptionsResult.Success;
-            }
-
-            return ValidateOptionsResult.Fail(result.ToString());
-        }
     }
 }
